Backend/requirements.txt
# Core Framework
#python version conda install python=3.12
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
pydantic==2.5.0
python-multipart==0.0.6

# Google Cloud Services
google-cloud-documentai==2.26.0
google-cloud-firestore==2.14.0
google-cloud-storage==2.13.0
google-cloud-secret-manager==2.18.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0

# JWT and Cryptography for Google Wallet
PyJWT==2.8.0
cryptography==41.0.7

# AI/ML
google-generativeai==0.3.2
langchain==0.1.0
langchain-google-genai==0.0.5

# Database
firebase-admin==6.4.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pillow==10.1.0
pydantic-settings==2.1.0

# Date/Time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# CORS
fastapi-cors==0.0.6

# Logging
structlog==23.2.0